Save logic moved to models. Parsers removed
