--- server/lib/parser/index.js
@@ -1,7 +1,0 @@
-const Block       = require('./block');
-const Transaction = require('./transaction');
-
-module.exports = {
-  Block,
-  Transaction,
-};
