+++ server/models/block.js
@@ -1,5 +1,6 @@
 const mongoose = require('mongoose');
 const config = require('../config');
+const util = require('../lib/util');
 
 const Schema = mongoose.Schema;
 // These limits can be overriden higher up the stack
@@ -64,6 +65,34 @@ BlockSchema.statics.getHeights = function findMissing(cb) {
     { height: 1 },
     cb)
     .sort({ height: 1 });
+};
+
+BlockSchema.statics.saveBcoinBlock = function saveBcoinBlock(entry, block, cb) {
+  const Block = this.model('Block');
+  const rawBlock = block.toRaw().toString('hex');
+  const blockJSON = block.toJSON();
+  const reward = util.calcBlockReward(entry.height);
+
+  return new Block({
+    hash: blockJSON.hash,
+    height: entry.height,
+    size: block.getSize(),
+    version: blockJSON.version,
+    prevBlock: blockJSON.prevBlock,
+    merkleRoot: blockJSON.merkleRoot,
+    ts: blockJSON.ts,
+    bits: blockJSON.bits,
+    nonce: blockJSON.nonce,
+    txs: block.txs.map((tx) => {
+      const txJSON = tx.toJSON();
+      return txJSON.hash;
+    }),
+    chainwork: entry.chainwork,
+    reward,
+    network: config.bcoin.network,
+    poolInfo: {},
+    rawBlock,
+  }).save(cb);
 };
 
 module.exports = mongoose.model('Block', BlockSchema);
